server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "cryptography>=46.0.6",
    "fastapi>=0.135.2",
    "greenlet>=3.3.2",
    "langchain-core>=1.2.23",
    "langchain-openai>=1.1.12",
    "langchain-xai>=1.2.2",
    "langgraph>=1.1.3",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]
